Api_whatsapp/contacto/serializers.py
from rest_framework import serializers
from contacto.models import Contacto


class ContactoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contacto
        fields = '__all__'

