Api_whatsapp/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
asyncpg==0.29.0
beautifulsoup4==4.12.2
blinker==1.8.2
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==41.0.3
cssselect2==0.7.0
distlib==0.3.7
Django==4.2.3
django-bootstrap4==23.2
django-cors-headers==4.2.0
django-crispy-forms==2.1
django-environ==0.11.2
djangorestframework==3.14.0
filelock==3.12.4
Flask==3.0.3
fonttools==4.42.0
html5lib==1.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==4.9.3
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
oscrypto==1.3.0
Pillow==10.0.0
platformdirs==3.11.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.7.0
pygame==2.5.2
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pypdf==3.15.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-decouple==3.8
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.4
virtualenv==20.24.5
webencodings==0.5.1
Werkzeug==3.0.3
xhtml2pdf==0.2.11
zopfli==0.2.2
